+++ src/components/DetailsPanel/DetailsPanel.tsx
