+++ src/components/DetailsPanel/index.ts
@@ -1,0 +1,1 @@
+export { default as DetailsPanel } from "./DetailsPanel";
