+++ src/components/NavBar/NavBar.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import styled from "styled-components";
 import { Link, useLocation } from "react-router-dom";
 
-import Button from "./Button";
-import Icon from "./Icon";
-import { NavigationLink } from "../types/NavigationLink";
+import { Button } from "../Button";
+import { Icon } from "../Icon";
+import { NavigationLink } from "../../types/NavigationLink";
 
 const NavBarWrapper = styled.nav`
   display: flex;
