+++ src/components/TopBar/TopBar.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 
-import Button from "./Button";
-import Icon from "./Icon";
+import { Button } from "../Button";
+import { Icon } from "../Icon";
 
 const ToolBarWrapper = styled.div`
   font-size: ${props => props.theme.fontSizes.text3};
