+++ src/components/TopBar/index.ts
@@ -1,0 +1,1 @@
+export { default as TopBar } from "./TopBar";
