+++ src/pages/UnderConstruction/index.ts
@@ -1,0 +1,1 @@
+export { default as UnderConstruction } from "./UnderConstruction";
