+++ src/tests/components/NavBar.test.tsx
@@ -2,7 +2,7 @@ import { screen } from "@testing-library/react";
 import React from "react";
 import { MemoryRouter } from "react-router-dom";
 
-import NavBar from "../../components/NavBar";
+import { NavBar } from "../../components/NavBar";
 import { themedRender } from "../helpers/themedRender";
 import MainTheme from "../../assets/themes/MainTheme";
